feat: use global store, fix all re-rendering issues
