+++ frontend/src/components/base/radiogroup/index.ts
@@ -1,0 +1,1 @@
+export * from "./base";
