+++ frontend/src/dashboard/components/index.ts
@@ -2,5 +2,4 @@ export * from "./MatchDialog";
 export * from "./QuestionRadioGroup";
 export * from "./RoomEditor";
 export * from "./RoomListBox";
-export * from "./LeaveRoomButton";
 export * from "./LoadedRoom";
