+++ frontend/src/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from "./useAuth";
 export * from "./useMobile";
 export * from "./useScrollDirection";
