--- frontend/src/room/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./useEditor";
